fix billing lint checks
